Introduce SetIfPresent and move metrics out to new file
